misc: addressed comments and doc changes
